+++ Fractol/libft/srcs/ft_lerp.c
@@ -6,13 +6,13 @@
 /*   By: saneveu <saneveu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/25 02:46:59 by saneveu           #+#    #+#             */
-/*   Updated: 2019/05/25 02:49:24 by saneveu          ###   ########.fr       */
+/*   Updated: 2019/05/28 04:48:21 by saneveu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double      ft_lerp(double first, double second, int p)
+double      ft_lerp(double first, double second, double p)
 {
     if (first == second)
         return (first);
