+++ Fractol/libft/srcs/ft_lerpi.c
@@ -6,13 +6,13 @@
 /*   By: saneveu <saneveu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/25 02:49:33 by saneveu           #+#    #+#             */
-/*   Updated: 2019/05/25 02:51:33 by saneveu          ###   ########.fr       */
+/*   Updated: 2019/05/28 04:47:06 by saneveu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int      ft_lerpi(int first, int second, int p)
+int      ft_lerpi(int first, int second, double p)
 {
     if (first == second)
         return (first);
